rename connection util to localStorage
